moving the ApiMessage to the message module
