+++ matrix/services/message/src/main/kotlin/app/dapk/st/matrix/message/internal/ApiMessage.kt
@@ -1,0 +1,26 @@
+package app.dapk.st.matrix.message.internal
+
+import app.dapk.st.matrix.common.MessageType
+import app.dapk.st.matrix.common.RoomId
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+sealed class ApiMessage {
+
+    @Serializable
+    @SerialName("text_message")
+    data class TextMessage(
+        @SerialName("content") val content: TextContent,
+        @SerialName("room_id") val roomId: RoomId,
+        @SerialName("type") val type: String,
+    ) : ApiMessage() {
+
+        @Serializable
+        data class TextContent(
+            @SerialName("body") val body: String,
+            @SerialName("msgtype") val type: String = MessageType.TEXT.value,
+        )
+    }
+
+}
